RSPromoBot/requirements.txt
aiohttp>=3.8.0
discord.py>=2.5.2
python-dotenv>=1.0.1
